requirements.txt
langgraph>=0.0.10
langchain>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymongo>=4.0.0
redis>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
loguru>=0.7.0 